rename readonly prop to disabled
