+++ src/lib/MultiSelect.svelte
@@ -3,7 +3,7 @@
   import { fly } from 'svelte/transition'
   import type { Option, Primitive, ProtoOption, DispatchEvents } from './'
   import CircleSpinner from './CircleSpinner.svelte'
-  import { CrossIcon, ExpandIcon, ReadOnlyIcon } from './icons'
+  import { CrossIcon, ExpandIcon, DisabledIcon } from './icons'
   import Wiggle from './Wiggle.svelte'
 
   export let selected: Option[] = []
@@ -13,7 +13,7 @@
   export let showOptions = false
   export let maxSelect: number | null = null // null means any number of options are selectable
   export let maxSelectMsg: ((current: number, max: number) => string) | null = null
-  export let readonly = false
+  export let disabled = false
   export let options: ProtoOption[]
   export let input: HTMLInputElement | null = null
   export let placeholder: string | undefined = undefined
@@ -103,7 +103,7 @@
   function add(label: Primitive) {
     if (maxSelect && maxSelect > 1 && selected.length >= maxSelect) wiggle = true
     if (
-      !readonly &&
+      !disabled &&
       !selectedLabels.includes(label) &&
       // for maxselect = 1 we always replace current option with new selection
       (maxSelect === null || maxSelect === 1 || selected.length < maxSelect)
@@ -126,7 +126,7 @@
   }
 
   function remove(label: Primitive) {
-    if (selected.length === 0 || readonly) return
+    if (selected.length === 0 || disabled) return
     const option = _options.find((option) => option.label === label)
     if (!option) {
       return console.error(`MultiSelect: option with label ${label} not found`)
@@ -221,7 +221,7 @@
 <!-- z-index: 2 when showOptions is true ensures the ul.selected of one <MultiSelect />
 display above those of another following shortly after it -->
 <div
-  class:readonly
+  class:disabled
   class:single={maxSelect === 1}
   class:open={showOptions}
   class="multiselect {outerDivClass}"
@@ -240,7 +240,7 @@ display above those of another following shortly after it -->
         <slot name="selected" {option} {idx}>
           {option.label}
         </slot>
-        {#if !readonly}
+        {#if !disabled}
           <button
             on:mouseup|stopPropagation={() => remove(option.label)}
             on:keydown={handleEnterAndSpaceKeys(() => remove(option.label))}
@@ -272,8 +272,8 @@ display above those of another following shortly after it -->
       <CircleSpinner />
     </slot>
   {/if}
-  {#if readonly}
-    <ReadOnlyIcon height="14pt" />
+  {#if disabled}
+    <DisabledIcon height="14pt" />
   {:else if selected.length > 0}
     {#if maxSelect && (maxSelect > 1 || maxSelectMsg)}
       <Wiggle bind:wiggle angle={20}>
@@ -360,8 +360,8 @@ display above those of another following shortly after it -->
   :where(div.multiselect:focus-within) {
     border: var(--sms-focus-border, 1pt solid var(--sms-active-color, cornflowerblue));
   }
-  :where(div.multiselect.readonly) {
-    background: var(--sms-readonly-bg, lightgray);
+  :where(div.multiselect.disabled) {
+    background: var(--sms-disabled-bg, lightgray);
   }
 
   :where(div.multiselect > ul.selected) {
