+++ src/lib/icons/Disabled.svelte
