+++ src/lib/icons/index.ts
@@ -1,3 +1,3 @@
-export { default as CrossIcon } from './Cross.svelte'
 export { default as ExpandIcon } from './ChevronExpand.svelte'
-export { default as ReadOnlyIcon } from './ReadOnly.svelte'
+export { default as CrossIcon } from './Cross.svelte'
+export { default as DisabledIcon } from './Disabled.svelte'
